# Conflicts:
#	src/main/java/com/tre3p/fileserver/service/impl/FileServiceImpl.java

+++ src/main/java/com/tre3p/fileserver/service/impl/FileServiceImpl.java
@@ -27,12 +27,12 @@ public class FileServiceImpl implements FileService {
     private static final String DATASTORAGE = "/datastorage/";
 
     @Override
-    public List<FileMetadata> getAll() {
+    public final List<FileMetadata> getAll() {
         return fileRepository.findAll();
     }
 
     @Override
-    public void removeById(Integer id) throws FileNotFoundException {
+    public final void removeById(Integer id) throws FileNotFoundException {
         log.info("+removeById(): id: {}", id);
         FileMetadata savedFile = fileRepository.findById(id)
                         .orElseThrow(() -> new FileNotFoundException("File not exists"));
@@ -48,7 +48,7 @@ public class FileServiceImpl implements FileService {
     }
 
     @Override
-    public FileMetadata prepareForSaving(String fileName, String contentType, File file) throws IOException {
+    public final FileMetadata prepareForSaving(String fileName, String contentType, File file) throws IOException {
         log.info("+prepareAndSave(): filename: {}, contentType: {}", fileName, contentType);
         if (file.exists()) {
             log.info("prepareAndSave(): file exists, saving..");
@@ -59,13 +59,11 @@ public class FileServiceImpl implements FileService {
 
             String originalSize = calculateSize(newFile.length());
 
-            int fileFormatNum = compareFileSizes(newFile.length(), zippedFile.length());
-            File fileForSave = fileFormatNum == 1 ? zippedFile : newFile;
-            deleteFileFormat(fileFormatNum, newFile, zippedFile);
+            File fileForSave = compareFileSizes(newFile, zippedFile);
             String zippedSize = calculateSize(zippedFile.length());
 
             log.info("-prepareAndSave(): file successfully saved at {}", fileForSave.getPath());
-            return save(fileName, contentType, fileForSave, fileFormatNum, originalSize, zippedSize);
+            return save(fileName, contentType, fileForSave, originalSize, zippedSize);
         } else {
             log.error("-prepareAndSave(): file not exists");
             throw new FileNotFoundException("File not exists");
@@ -73,18 +71,21 @@ public class FileServiceImpl implements FileService {
     }
 
     @Override
-    public FileMetadata getById(Integer id) {
+    public final FileMetadata getById(Integer id) {
         return fileRepository.getById(id);
     }
 
     @Override
-    public FileMetadata save(String fileName, String contentType, File file,
-                             int formatNum, String originalSize, String zippedSize) {
+    public final FileMetadata save(String fileName,
+                                   String contentType,
+                                   File file,
+                                   String originalSize,
+                                   String zippedSize) {
         return fileRepository.save(new FileMetadata(
                 fileName,
                 file.getName(),
                 contentType,
-                isZipped(formatNum),
+                false,
                 originalSize,
                 zippedSize,
                 file.getAbsolutePath()
@@ -97,27 +98,19 @@ public class FileServiceImpl implements FileService {
         return result.equals("0 bytes") ? "-" : result; // todo: костыль, надо переделать
     }
 
-    private int compareFileSizes(long originalSize, long zippedSize) {
-        log.info("+compareFileSizes(): originalSize: {}. zippedSize: {}", originalSize, zippedSize);
-        int result =  zippedSize > originalSize ? 0 : 1;
-        log.info("compareFileSizes(): result {}", result);
-        return result;
-    }
-
-    private void deleteFileFormat(int fileFormatNum, File originalFile, File zippedFile) { // todo: сделать красивей
-        if (fileFormatNum == 1) {
-            log.info("deleteFileFormat: deleting original file {} ...", originalFile);
-            originalFile.delete();
-            return;
+    private File compareFileSizes(File nativeFile, File zippedFile) {
+        long originalSize = nativeFile.length();
+        long zippedSize = zippedFile.length();
+        log.info("compareFileSizes(): originalSize: {}. zippedSize: {}", originalSize, zippedSize);
+        if (zippedSize > originalSize) {
+            zippedFile.delete(); // todo: костыль поправить надо. отсюда убрать удаление файла и возвращать 0/1
+            log.info("-compareFileSizes(): file saved as native format file");
+            return nativeFile;
         }
-        log.info("deleteFileFormat: deleting zipped file {} ...", zippedFile);
-        zippedFile.delete();
+        nativeFile.delete();
+        log.info("-compareFileSizes(): file saved as zipped format file");
+        return zippedFile;
     }
-
-    private boolean isZipped(int formatNum) {
-        return formatNum == 1 ? true : false; // todo: перенести метод в зип сервис
-    }
-
 
     private File moveFileToMainStorage(File file, String originalFileName) throws IOException {
         return Files.move(Paths.get(
